pyproject.toml
[project]
name = "mcp-alchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]===1.2.0rc1",
    "psycopg2-binary>=2.9.10",
    "pymysql>=1.1.1",
    "sqlalchemy>=2.0.36",
    "azure-identity>=1.15.0",
    "azure-core>=1.30.0",
]
